Split up Twilio REST client objects
